gameplay: seed the lane table, wake the pause state machine, show the bursts
Three separate faults, each with a single root cause.

NOTES ALL LANDED ON THE CENTRE LINE. `InitNoteLaneTable` was never invoked.
0x136afc is a `__mod_init_func` entry — its pointer sits at 0x358cb8, and
`get_xrefs_to` finds no call site anywhere in the binary, only that DATA
reference — so dyld runs it at image load. The reconstruction had neither the
constructor attribute nor a call, so `g_noteLaneTable` stayed zero-initialised,
`GetNoteLaneFraction` returned 0.0f for every lane and kind, and both a note's
base X and its route's final X were 0. Every note therefore spawned on, ran
down, and landed on the horizontal centre. The table's values were already
correct — they were simply never written, which is why reading the source made
it look right.

`NoteModel::Init`'s stack lane-fraction lookup was five entries where the binary
builds ten (the seven ordinary fractions plus three trailing zeros, written at
0x134204-0x134250). It is indexed by the note's colour, which runs 0-6, and by a
slide point's target, which runs 0-9, so lanes 5 and 6 and six of the ten slide
targets read past the end of the array.

PAUSING LOCKED THE GAME. The pause scene's per-frame state machine was left as a
non-virtual `TaskExecute` that nothing called, while an unrelated empty stub at
0x1508b0 was bound to the `OnFrame` override. The binary's vtable settles it:
the constructor at 0x1508b4 installs vptr 0x35dad0, whose first slot is 0x150b38
— the state machine. So the listener dispatch runs the state machine every
frame, and with it bound to the empty stub the machine never advanced past its
load state: the menu never opened and nothing could resume. The state machine is
now the override; the stub is gone.

Two more on that path: the Limelight/Colette themes dim the Exit lane *while* a
pastel bonus is active (matching `HandleExit`, which refuses to exit under
exactly that condition), not when one is absent; and `ExecShow` falls straight
through from a successful touch acquisition into the drag block on the same
frame, which an `else` prevented.

NO EXPLOSIONS ON A HIT. `ExplosionEffectLayer`'s constructor dropped three seeds
the binary performs: both play-colour alpha bytes to 0xff (0x176e4c-0x176e54),
the effect size to 1.0f (0x176e58), and both banks' effect type to the disabled
value via the eight-byte `_memset_pattern16` at 0x176e7c. `Process` reads the
alpha byte as every burst's sprite alpha, so leaving it zero made each burst
fully transparent — created, animated, and invisible.

Also corrects `ReserveNoteLane`'s range guard to the binary's unsigned compare
(0x14918c), so a hold note with colour tone -1 no longer writes outside the slot
array.

+++ Project/OpenGL/Models/note_lane_tracker.mm
@@ -131,8 +131,9 @@ int NoteLaneTracker::AssignNoteLane(
 /** @ghidraAddress 0x149178 */
 void NoteLaneTracker::ReserveNoteLane(
     int nTimeStart, int nDuration, int nPlayer, int nLane, bool bSpread) {
-    // Only the first three lane groups are reservable.
-    if (nLane > 2) {
+    // Only the first three lane groups are reservable; the binary's test is unsigned, so a
+    // negative colour tone is rejected here too (the b.hi at 0x14918c).
+    if (static_cast<unsigned int>(nLane) > 2) {
         return;
     }
     const int nTimeEnd = nDuration + nTimeStart;
